APP/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psycopg2-binary==2.9.5
pytest==7.4.2
flasgger==0.9.7.1
flask-cors==4.0.0
python-dotenv==1.0.0
